+++ src/server/game/Globals/ObjectMgr.cpp
@@ -4414,8 +4414,8 @@ void ObjectMgr::LoadPlayerInfo()
             }
 
             PlayerClassLevelInfo& levelInfo = info->levelInfo[current_level - 1];
-            levelInfo.basehealth = fields[2].GetUInt16();
-            levelInfo.basemana   = fields[3].GetUInt16();
+            levelInfo.basehealth = fields[2].GetUInt32();
+            levelInfo.basemana   = fields[3].GetUInt32();
 
             ++count;
         }
@@ -4504,8 +4504,8 @@ void ObjectMgr::LoadPlayerInfo()
                     info->levelInfo = std::make_unique<PlayerLevelInfo[]>(sWorld->getIntConfig(CONFIG_MAX_PLAYER_LEVEL));
 
                 PlayerLevelInfo& levelInfo = info->levelInfo[current_level - 1];
-                for (uint8 i = 0; i < MAX_STATS; ++i)
-                    levelInfo.stats[i] = fields[i + 3].GetUInt8();
+                for (uint16 i = 0; i < MAX_STATS; ++i)
+                    levelInfo.stats[i] = fields[i + 3].GetUInt16();
             }
 
             ++count;
